Moved FinnHub to a folder inside of Infrastructure
