+++ StonksBackend/Domain/Entities/FinnHubResolution.cs
@@ -1,0 +1,14 @@
+namespace StonksBackend.Infrastructure
+{
+    public enum FinnHubResolution
+    {
+        ONE = 1,
+        FIVE = 5,
+        FIFTEEN = 15,
+        THIRTY = 30,
+        SIXTY = 60,
+        D,
+        W,
+        M
+    }
+}
